refactor : cars 라는 일급컬렉션을 만듬 1. CarFactory 객체제거 및 팩터리 메서드 Cars로 이관 2. RacingResult 에 있던 우승자 찾는 메서드 Cars로 이관
